requirements.txt
flask==2.3.3
flask-cors==4.0.0
onnxruntime==1.24.1
numpy==1.24.4
pillow==10.2.0
werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
gdown==5.1.0
